Add confirmation request to Init.sh

+++ Init.sh
@@ -10,8 +10,8 @@ if [ "$1" == "-h" ] || [ "$1" == "-help" ] || [ "$1" == "--help" ] || [ "$1" == 
     echo " - Create links to Hog git-hooks"
     echo " - Locally ignore Xilinx IP xml files"
     echo " - Initialise and update your submodules"
-    echo " - Create vivado projects (if vivado exacutable is found)"
-    echo " - Compile questasim libraries (if questasim executable is found)"
+    echo " - (optional) Compile questasim libraries (if questasim executable is found)"
+    echo " - (optional) Create vivado projects (if vivado exacutable is found)"
     echo
     exit 0
 fi
@@ -46,24 +46,35 @@ then
     VIVADO=`which vivado`
     if [ `which vsim` ]
     then
-	echo [hog init] Compiling Modelsim libraries into ../ModelsimLib...
-	$VIVADO -mode batch -notrace -source ./Tcl/compile_library.tcl
-	rm -f ./Tcl/.cxl.questasim.version
-	rm -f ./Tcl/compile_simlib.log
-	rm -f ./Tcl/modelsim.ini
+	echo
+	read -p "Do you want to compile Questasim libraries (this might take some time)? " -n 1 -r
+	echo  
+	if [[  $REPLY =~ ^[Yy]$ ]]
+	then
+	    echo [hog init] Compiling Modelsim libraries into ../ModelsimLib...
+	    $VIVADO -mode batch -notrace -source ./Tcl/compile_library.tcl
+	    rm -f ./Tcl/.cxl.questasim.version
+	    rm -f ./Tcl/compile_simlib.log
+	    rm -f ./Tcl/modelsim.ini
+	fi
     else
 	echo [hog init] "WARNING: No modelsim executable found, will not compile libraries"
     fi
-
-    cd ../Top
-    proj=`ls`
-    cd ..
-    echo [hog init] Creating projects for: $proj...
-    for f in $proj
-    do
-	echo [hog init] Creating Vivado project: $f...
-	./Hog/CreateProject.sh $f
-    done
+    echo
+    read -p "Do you want to create projects now (can be done later with CreateProject.sh)? " -n 1 -r
+    echo    # (optional) move to a new line
+    if [[ $REPLY =~ ^[Yy]$ ]]
+    then
+	cd ../Top
+	proj=`ls`
+	cd ..
+	echo [hog init] Creating projects for: $proj...
+	for f in $proj
+	do
+	    echo [hog init] Creating Vivado project: $f...
+	    ./Hog/CreateProject.sh $f
+	done
+    fi
 else
     echo [hog init] "WARNING: No vivado executable found"
 fi
